- fixed lemona_blade_string_null to incorporate NULL character. This   should avoid a buffer overflow when trying to copy the string to the   zest - added missing headers - fixed conditions used to avoid logging action on ours relayfs files. - added stub for lemona_relay's functions to avoid compile errors if   disabled - fixed a bug in lemona_zest_fill causing arguments size to be overriden - fixed lemona_blade_string_null, the \0 wasn't accounted in the arg size - updated lemona_net_log to loop on kernel_sendmsg until all data are sent - added another quick & dirty tool to visualize content of log. It is   uncomplete and doesn't handle truncated records (yet); itllowed me   to check that basket was correctly receiving & storing the zests   though. - modified basket to use low level write instead of mmap (for a reason   data were lost, didn't have the time to look into it too long). It   is still a quick & dirty implementation though.
- TODO:
  - find why there is this hangup when system halt while lemona_net
  was activated and connected to a server (How to detect that the
  interface was shutdown?)

+++ driver/blades/generics.c
@@ -132,7 +132,7 @@ int		lemona_blade_output_buffer(struct lemona_zest *zest,
 					   int isExt, int idx, int off,
 					   void __user *buf, void *len)
 {
-  ssize_t	size	= *((int *)len);
+  ssize_t	size	= *((ssize_t *)len);
   int		*sz;
   unsigned long	uncopied = size;
 
@@ -143,7 +143,7 @@ int		lemona_blade_output_buffer(struct lemona_zest *zest,
   /* fill the zest */
   sz = isExt == false ? zest->argsz : zest->extsz;
   sz[idx] = sizeof(size);
-  *((int *)((char *)zest + off)) = size;
+  *((ssize_t *)((char *)zest + off)) = size;
 
   /* try to copy only if we actually got something to copy */
   if (size > 0)
